% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{certificate_df}
\alias{certificate_df}
\title{Generate a data frame with certificate information}
\usage{
certificate_df(x, database_true = isTruthy(config$mongo_key))
}
\arguments{
\item{x}{A list containing `data_formatted` and `rules` elements.}

\item{database_true}{A logical value (default: `isTruthy(config$mongo_key)`). If TRUE, the generated
data frame will be inserted into the database.}
}
\value{
A data frame with certificate information.
}
\description{
This function creates a data frame with certificate information including the current time,
data and rule hashes, package version, and web hash.
}
