% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{check_exists_in_zip}
\alias{check_exists_in_zip}
\title{Check if a file exists in a zip file}
\usage{
check_exists_in_zip(zip_path, file_name)
}
\arguments{
\item{zip_path}{A character string representing the path of the zip file.}

\item{file_name}{A character string representing the name of the file to check.}
}
\value{
A logical value indicating whether the file exists in the zip file (TRUE) or not (FALSE).
}
\description{
This function checks if a file with a given name exists in a specified zip file.
}
\examples{
\dontrun{
check_exists_in_zip(zip_path = "/path/to/your.zip", file_name = "file/in/zip.csv")
}
}
