% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{check_images}
\alias{check_images}
\title{Check and format image URLs}
\usage{
check_images(x)
}
\arguments{
\item{x}{A character string to check for image URLs.}
}
\value{
A character string with the HTML img tag if an image URL is found, otherwise the input string.
}
\description{
This function checks if the input string contains an image URL (PNG or JPG) and formats it as an HTML img tag with a specified height.
}
\examples{
check_images("https://example.com/image.png")
check_images("https://example.com/image.jpg")
check_images("https://example.com/text")
}
