% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{check_uploadable}
\alias{check_uploadable}
\title{Check if a file can be uploaded to an S3 bucket}
\usage{
check_uploadable(url, bucket = config$s3_bucket)
}
\arguments{
\item{url}{A character string representing the URL of the file to download.}

\item{bucket}{A character string representing the S3 bucket name where the file will be uploaded. Defaults to 'config$s3_bucket'.}
}
\value{
A logical value indicating whether the file can be uploaded (TRUE) or not (FALSE).
}
\description{
This function checks if a file located at a given URL can be downloaded and uploaded to an S3 bucket.
}
\examples{
# Note: The example assumes you have valid AWS credentials and an S3 bucket available
check_uploadable("https://example.com/file.csv", bucket = "your-s3-bucket-name")
}
