% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{checkLuhn}
\alias{checkLuhn}
\title{Check if a number passes the Luhn algorithm}
\usage{
checkLuhn(number)
}
\arguments{
\item{number}{A character string of the number to check against the Luhn algorithm.}
}
\value{
A logical value indicating whether the number passes the Luhn algorithm (TRUE) or not (FALSE).
}
\description{
This function checks if a given number passes the Luhn algorithm. It is commonly used to validate credit card numbers.
}
\examples{
checkLuhn("4532015112830366") # TRUE
checkLuhn("4532015112830367") # FALSE
}
