% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets2.R
\docType{data}
\name{invalid_example}
\alias{invalid_example}
\title{Invalid Example Data}
\format{
A list with 3 data frames:
\describe{
  \item{data_frame_1}{A data frame with 18 variables: MethodologyID, SamplingDevice, AirFiltration, AirFiltrationType, ClothingPolicy, NonplasticPolicy, SealedEnvironment, SealedEnvironmentType, SieveMeshSizes, FilterType, FilterDiameter, FilterPoreSize, VisIDMethod, VisualSoftware, PickingStrategy, VisMagnification, MatIDMethod, MatIDSoftware}
  \item{data_frame_2}{A data frame with 8 variables: SampleID, OwnerOrganization, AnalysisOrganization, ReportingOrganization, Latitude, Longitude, CollectionDate, SampleVolume}
  \item{data_frame_3}{A data frame with 17 variables: ParticleID, MethodologyID, SampleID, PhotoID, SpectraID, FinalAnalysisDate, Comments, Polymer, Morphology, Color, Length, Width, Height, Mass, SurfaceArea, Volume, Tactile}
}
}
\description{
This is a list containing three data frames as an example of invalid_example.
}
\examples{
data("invalid_example")

}
\keyword{datasets}
