% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{is.POSIXct}
\alias{is.POSIXct}
\title{Check if an object is of class POSIXct}
\usage{
is.POSIXct(x)
}
\arguments{
\item{x}{An object to be tested for POSIXct class inheritance.}
}
\value{
A logical value indicating if the input object is of class POSIXct.
}
\description{
This function checks if the given object is of class POSIXct.
It returns TRUE if the object inherits the POSIXct class, otherwise FALSE.
}
\examples{
x <- as.POSIXct("2021-01-01")
is.POSIXct(x) # TRUE

y <- Sys.Date()
is.POSIXct(y) # FALSE

}
