% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{name_data}
\alias{name_data}
\title{Name datasets}
\usage{
name_data(files_data, data_names = NULL)
}
\arguments{
\item{files_data}{A vector of file paths or list of data frames.}

\item{data_names}{A vector of names to be assigned to datasets.}
}
\value{
A vector of dataset names.
}
\description{
This function extracts the names of the datasets provided in the input files.
If specific data names are provided, they are used, otherwise the function tries
to extract the names from the files themselves.
}
\examples{
name_data(files_data = c("path/to/data1.csv", "path/to/data2.csv"))
name_data(files_data = c("path/to/data.xlsx"), data_names = c("sheet1", "sheet2"))

}
