% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{query_mongodb_api}
\alias{query_mongodb_api}
\title{Query MongoDB Document by ObjectID}
\usage{
query_mongodb_api(collection, database, dataSource, apiKey, objectId)
}
\arguments{
\item{collection}{The name of the collection in the MongoDB database.}

\item{database}{The name of the MongoDB database.}

\item{dataSource}{The data source in MongoDB.}

\item{apiKey}{The API key for accessing the MongoDB API.}

\item{objectId}{The object ID of the document to query.}
}
\value{
The queried document.
}
\description{
This function queries a mongodb database using its API to retrieve a document by its ObjectID.
}
\examples{
\dontrun{
result <- query_document_by_object_id(
        apiKey = 'your_mongodb_api_key',
        collection = 'your_mongodb_collection',
        database = 'your_database',
        dataSource = 'your_dataSource',
        objectId = 'example_object_id'
    )
}
}