% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{read_data}
\alias{read_data}
\title{Read and format data from csv or xlsx files}
\usage{
read_data(files_data, data_names = NULL)
}
\arguments{
\item{files_data}{List of files to be read}

\item{data_names}{Optional vector of names for the data frames}
}
\value{
A list of data frames
}
\description{
Read and format data from csv or xlsx files
}
\examples{
read_data(files_data = valid_example, data_names = c("methodology", "particles", "samples"))
}

