% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{read_rules}
\alias{read_rules}
\title{Read rules from a file}
\usage{
read_rules(file_rules)
}
\arguments{
\item{file_rules}{The file containing the rules. Can be a CSV or XLSX file, or a data frame.}
}
\value{
A data frame containing the rules.
}
\description{
This function reads rules from a file or a data frame.
The file can be in csv or xlsx format.
The data should have the column names "name", "description", "dataset", "valid example", "severity", "rule".
The function also checks that the rules do not contain sensitive words and that
all the rules fields are character type.
}
\examples{
\dontrun{
read_rules("path/to/rules")
}

}
