% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{reformat_rules}
\alias{reformat_rules}
\title{Reformats the rules}
\usage{
reformat_rules(rules, data_formatted)
}
\arguments{
\item{rules}{A data.frame containing rules to be reformatted.}

\item{data_formatted}{A named list of data.frames with data.}
}
\value{
A data.frame with reformatted rules.
}
\description{
This function is responsible for handling the rule reformating, dataset handling
and foreign key checks.
}
\examples{
reformat_rules(rules, data_formatted)
}
