% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{remote_share}
\alias{remote_share}
\title{Remote Share Function}
\usage{
remote_share(
  validation,
  data_formatted,
  verified,
  valid_rules,
  valid_key,
  ckan_url,
  ckan_key,
  ckan_package,
  url_to_send,
  rules,
  results,
  s3_key_id,
  s3_secret_key,
  s3_region,
  s3_bucket,
  mongo_key,
  old_cert = NULL
)
}
\arguments{
\item{validation}{A list containing validation information.}

\item{data_formatted}{A list containing formatted data.}

\item{verified}{The secret key provided by the portal maintainer.}

\item{valid_rules}{A list of valid rules for the dataset.}

\item{valid_key}{A list of valid keys.}

\item{ckan_url}{The URL of the CKAN instance.}

\item{ckan_key}{The API key for the CKAN instance.}

\item{ckan_package}{The CKAN package to which the data will be uploaded.}

\item{url_to_send}{The URL to send the data.}

\item{rules}{A set of rules used for validation.}

\item{results}{A list containing results of the validation.}

\item{s3_key_id}{AWS ACCESS KEY ID}

\item{s3_secret_key}{AWS SECRET ACCESS KEY}

\item{s3_region}{AWS DEFAULT REGION}

\item{s3_bucket}{The name of the Amazon S3 bucket.}

\item{mongo_key}{The API key for the MongoDB instance.}

\item{old_cert}{(Optional) An old certificate to be uploaded alongside the new one.}
}
\value{
A list containing the status and message of the operation.
}
\description{
This function uploads validated data to specified remote repositories,
such as CKAN, MongoDB, and/or Amazon S3.
}
\examples{
#Need to add. 

}
