% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{run_app}
\alias{run_app}
\title{Run any of the apps}
\usage{
run_app(
  path = "system",
  log = TRUE,
  ref = "main",
  test_mode = FALSE,
  app = "validator",
  ...
)
}
\arguments{
\item{path}{to store the downloaded app files; defaults to \code{"system"}
pointing to \code{system.file(package = "One4All")}.}

\item{log}{logical; enables/disables logging to \code{\link[base]{tempdir}()}}

\item{ref}{git reference; could be a commit, tag, or branch name. Defaults to
"main". Only change this in case of errors.}

\item{test_mode}{logical; for internal testing only.}

\item{app}{your app choice}

\item{\dots}{arguments passed to \code{\link[shiny]{runApp}()}.}
}
\value{
This function normally does not return any value, see
\code{\link[shiny]{runGitHub}()}.
}
\description{
This wrapper function starts the user interface of your app choice.
}
\details{
After running this function the Validator, Microplastic Image Explorer, or Data Visualization GUI should open in a separate
window or in your computer browser.
}
\examples{
\dontrun{
run_app(app = "validator")
}

}
\seealso{
\code{\link[shiny]{runGitHub}()}
}
\author{
Hannah Sherrod, Nick Leong, Hannah Hapich, Fabian Gomez, Win Cowger
}
