% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{test_rules}
\alias{test_rules}
\title{Rules data}
\format{
A data frame with 6 columns:
\describe{
  \item{name}{Name of the rule (e.g., "MethodologyID_valid")}
  \item{description}{Description of the rule (e.g., "URL address is valid and can be found on the internet.")}
  \item{dataset}{Dataset associated with the rule (e.g., "methodology")}
  \item{valid_example}{A valid example of the rule (e.g., "https://www.waterboards.ca.gov/drinking_water/certlic/drinkingwater/documents/microplastics/mcrplsts_plcy_drft.pdf")}
  \item{severity}{Severity of the rule (e.g., "error")}
  \item{rule}{The actual rule (e.g., "check_uploadable(MethodologyID) == TRUE")}
}
}
\description{
A dataset containing rules and their descriptions, datasets, valid examples, severity, and rules.
}
\examples{
data("test_rules")

}
\keyword{data}
