% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{query_document_by_object_id}
\alias{query_document_by_object_id}
\title{Query a MongoDB document by an ObjectID}
\usage{
query_document_by_object_id(apiKey, collection, database, dataSource, objectId)
}
\arguments{
\item{apiKey}{The API key for accessing the MongoDB API.}

\item{collection}{The name of the collection in the MongoDB database.}

\item{database}{The name of the MongoDB database.}

\item{dataSource}{The data source in MongoDB.}

\item{objectId}{The object ID of the document to query.}
}
\value{
The queried document.
}
\description{
This function queries a mongodb database using its API to retrieve a document by its ObjectID. Use the MongoDB Atlas Data API to create an API key.
}
\examples{
\dontrun{
apiKey <- 'your_mongodb_api_key'
collection <- 'your_mongodb_collection'
database <- 'your_database'
dataSource <- 'your_dataSource'
objectId <- 'example_object_id'
query_document_by_object_id(apiKey, collection, database, dataSource, objectId)
}

}
