% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{reformat_rules}
\alias{reformat_rules}
\title{Reformat the rules}
\usage{
reformat_rules(rules, data_formatted, zip_data = NULL)
}
\arguments{
\item{rules}{A data.frame containing rules to be reformatted.}

\item{data_formatted}{A named list of data.frames with data.}

\item{zip_data}{A file path to a zip folder with additional data to check.}
}
\value{
A data.frame with reformatted rules.
}
\description{
This function is responsible for handling the rule reformating, dataset handling
and foreign key checks.
}
\examples{
data("test_rules")
data("valid_example")
reformat_rules(rules = test_rules, data_formatted = valid_example)
}
